# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegra",
    "description": "Production-ready Agent Protocol server",
    "version": "0.17.0"
  },
  "paths": {
...
    "/store/items": {
      "put": {
        "tags": [
          "Store"
        ],
        "summary": "Put Store Item",
        "description": "Create or update an item in the store.\n\nIf an item with the same namespace and key already exists, its value is\noverwritten. Values must be JSON objects (dictionaries).",
        "operationId": "put_store_item_store_items_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorePutRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Store"
        ],
        "summary": "Get Store Item",
        "description": "Get an item from the store by key.\n\nReturns 404 if no item exists at the given namespace and key. Pass\n`refresh_ttl` to extend the item's lifetime as a side effect of reading it.",
        "operationId": "get_store_item_store_items_get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Key of the item to retrieve.",
              "title": "Key"
            },
            "description": "Key of the item to retrieve."
          },
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Namespace path. Use dot-separated string or repeated query params.",
              "title": "Namespace"
            },
            "description": "Namespace path. Use dot-separated string or repeated query params."
          },
          {
            "name": "refresh_ttl",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Extend the item's TTL on read.",
              "title": "Refresh Ttl"
            },
            "description": "Extend the item's TTL on read."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Store"
        ],
        "summary": "Delete Store Item",
        "description": "Delete an item from the store.\n\nAccepts parameters via JSON body (`namespace` + `key`) or query\nparameters. The JSON body takes precedence when both are provided.",
        "operationId": "delete_store_item_store_items_delete",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Key of the item to delete (query param alternative).",
              "title": "Key"
            },
            "description": "Key of the item to delete (query param alternative)."
          },
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Namespace path (query param alternative).",
              "title": "Namespace"
            },
            "description": "Namespace path (query param alternative)."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StoreDeleteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/items/search": {
      "post": {
        "tags": [
          "Store"
        ],
        "summary": "Search Store Items",
        "description": "Search items in the store.\n\nFilter items by namespace prefix, key-value metadata filters, or semantic\nquery. Results are paginated via `limit` and `offset`. When `query` is given\neach item carries its relevance `score`.",
        "operationId": "search_store_items_store_items_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/store/namespaces": {
      "post": {
        "tags": [
          "Store"
        ],
        "summary": "List Namespaces",
        "description": "List namespaces in the store.\n\nReturns the namespace paths that contain items. Filter by prefix, suffix,\nor maximum depth.",
        "operationId": "list_namespaces_store_namespaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreListNamespacesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListNamespacesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentProtocolError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "AgentProtocolError",
        "description": "Standard Agent Protocol error response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StoreDeleteRequest": {
        "properties": {
          "namespace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Namespace",
            "description": "Namespace path of the item to delete."
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Key of the item to delete."
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "key"
        ],
        "title": "StoreDeleteRequest",
        "description": "Request body for deleting store items (SDK-compatible)."
      },
      "StoreGetResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The item's key within its namespace."
          },
          "value": {
            "title": "Value",
            "description": "The stored value."
          },
          "namespace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Namespace",
            "description": "The namespace path where this item is stored."
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "When the item was first written."
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "When the item was last written."
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "namespace"
        ],
        "title": "StoreGetResponse",
        "description": "Response model for getting items"
      },
      "StoreItem": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The item's key within its namespace."
          },
          "value": {
            "title": "Value",
            "description": "The stored value."
          },
          "namespace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Namespace",
            "description": "The namespace path where this item is stored."
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "When the item was first written."
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "When the item was last written."
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Semantic relevance; null unless `query` was given."
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "namespace"
        ],
        "title": "StoreItem",
        "description": "Store item model"
      },
      "StoreListNamespacesRequest": {
        "properties": {
          "prefix": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix",
            "description": "Filter by namespace prefix"
          },
          "suffix": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suffix",
            "description": "Filter by namespace suffix"
          },
          "max_depth": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Depth",
            "description": "Maximum namespace depth to return"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum results",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Results offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "StoreListNamespacesRequest",
        "description": "Request model for listing store namespaces"
      },
      "StoreListNamespacesResponse": {
        "properties": {
          "namespaces": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Namespaces"
          }
        },
        "type": "object",
        "required": [
          "namespaces"
        ],
        "title": "StoreListNamespacesResponse",
        "description": "Response model for listing store namespaces"
      },
      "StorePutRequest": {
        "properties": {
          "namespace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Namespace",
            "description": "Storage namespace"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Item key"
          },
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value",
            "description": "Item value (must be a JSON object)"
          },
          "index": {
            "anyOf": [
              {
                "type": "boolean",
                "const": false
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index",
            "description": "Fields to embed for semantic search, as dotted paths into `value`. `false` skips indexing this item; omit to use the store's configured default."
          },
          "ttl": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl",
            "description": "Lifetime in minutes; omit to keep the item indefinitely."
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "key",
          "value"
        ],
        "title": "StorePutRequest",
        "description": "Request model for storing items"
      },
      "StoreSearchRequest": {
        "properties": {
          "namespace_prefix": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Namespace Prefix",
            "description": "Namespace prefix to search"
          },
          "filter": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter",
            "description": "Optional dictionary of key-value pairs to filter results."
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Search query"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum results",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Results offset",
            "default": 0
          },
          "refresh_ttl": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Ttl",
            "description": "Extend the TTL of matched items; omit to use the store's configured default."
          }
        },
        "type": "object",
        "required": [
          "namespace_prefix"
        ],
        "title": "StoreSearchRequest",
        "description": "Request model for searching store items"
      },
      "StoreSearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoreItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "StoreSearchResponse",
        "description": "Response model for store search"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}